refactor(condo): DOMA-8523 fixed imports
